+++ back/src/main/java/com/seguridadlimite/models/dao/IArlDao.java
@@ -4,6 +4,6 @@ package com.seguridadlimite.models.dao;
 import com.seguridadlimite.models.entity.Arl;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IArlDao extends CrudRepository<Arl, Long>{
+public interface IArlDao extends CrudRepository<Arl, Integer> {
 
 }
